Upload & Download file

+++ controllers/admin.js
@@ -1,4 +1,7 @@
 const mongoose = require('mongoose');
+
+const fileHelper = require('../util/file');
+
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
 
@@ -15,12 +18,11 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file
     const price = req.body.price;
     const description = req.body.description;
-    const errors = validationResult(req);
-
-    if(!errors.isEmpty()) {
+    
+    if(!image) {
         return res.status(422).render('admin/edit-product', {
             pageTitle: 'Add Product',
             path: '/admin/edit-product',
@@ -28,7 +30,25 @@ exports.postAddProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image',
+            validationErrors: []
+        });
+    }
+
+    const errors = validationResult(req);
+
+    if(!errors.isEmpty()) {
+        console.log(errors)
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/edit-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
                 price: price,
                 description: description
             },
@@ -36,6 +56,8 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array()
         })
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({
         title: title,
@@ -89,7 +111,8 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
+    console.log(image);
     const updatedDescription = req.body.description;
 
     const errors = validationResult(req);
@@ -102,7 +125,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription,
                 _id: prodId
@@ -121,7 +143,10 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDescription;
-            product.imageUrl = updatedImageUrl;
+            if(image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
             return product.save()
                 .then(result => {
                     console.log('PRODUCT UPDATED!');
@@ -154,8 +179,15 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user.id })
-        .then(result => {
+    Product.findById(prodId)
+        .then(product => {
+            if(!product) {
+                return next(new Error('Product not found'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user.id })
+        })
+        .then(() => {
             console.log('PRODUCT DELETED !');
             res.redirect('/admin/products');
         })
